refactor(review): split hunk mapping into diff parser and platform adapters
mapping.ts mixed four jobs in one 333-line file: a stateful diff parser, the
in-diff anchor validation, the GitHub and GitLab payload mappers, and CLI
orchestration. Pull the parser into diff.ts (testable in isolation) and the
validation plus both platform adapters into platform.ts. mapping.ts is now CLI
glue: parse args, parse the diff, validate, dispatch to the right adapter.

Collapse the noteAnchor wrapper that only re-exported deriveAnchor. Split the
tests to match: diff.test.ts, platform.test.ts, note.test.ts (deriveAnchor plus
the new decodeNotes). CLI output payloads are unchanged.
